requirements.txt
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
cachetools==4.2.1
cairocffi==1.2.0
CairoSVG==2.5.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cssselect==1.1.0
cssselect2==0.4.1
curlify==2.2.1
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
expertai-nlapi==2.1.3
gitdb==4.0.5
GitPython==3.1.12
html5lib==1.1
idna==2.10
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
lxml==4.6.2
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.5
notebook==6.2.0
numpy==1.20.1
packaging==20.9
pandas==1.2.1
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.14
protobuf==3.14.0
ptyprocess==0.7.0
pyarrow==3.0.0
pycparser==2.20
pydeck==0.6.0
Pygments==2.7.4
pyparsing==2.4.7
Pyphen==0.10.0
pyquery==1.4.3
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.0.2
requests==2.25.1
Send2Trash==1.5.0
six==1.15.0
smmap==3.0.5
streamlit==0.76.0
streamlit-agraph==0.0.35
terminado==0.9.2
testpath==0.4.4
tinycss2==1.1.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
tzlocal==2.1
urllib3==1.26.3
validators==0.18.2
watchdog==1.0.2
wcwidth==0.2.5
WeasyPrint==52.2
webencodings==0.5.1
widgetsnbextension==3.5.1
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
cachetools==4.2.1
cairocffi==1.2.0
CairoSVG==2.5.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cssselect==1.1.0
cssselect2==0.4.1
curlify==2.2.1
decorator==4.4.2
defusedxml==0.6.0
docx2txt==0.8
entrypoints==0.3
expertai-nlapi==2.1.3
gitdb==4.0.5
GitPython==3.1.12
html5lib==1.1
idna==2.10
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
langid==1.1.6
lxml==4.6.2
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.5
notebook==6.2.0
numpy==1.20.1
packaging==20.9
pandas==1.2.1
pandocfilters==1.4.3
parso==0.8.1
pdfminer.six==20200517
pdfplumber==0.5.26
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.14
protobuf==3.14.0
ptyprocess==0.7.0
pyarrow==3.0.0
pycparser==2.20
pycryptodome==3.10.1
pydeck==0.6.0
Pygments==2.7.4
pyparsing==2.4.7
PyPDF2==1.26.0
Pyphen==0.10.0
pyquery==1.4.3
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.0.2
requests==2.25.1
Send2Trash==1.5.0
six==1.15.0
smmap==3.0.5
sortedcontainers==2.3.0
streamlit==0.76.0
streamlit-agraph==0.0.35
terminado==0.9.2
testpath==0.4.4
tinycss2==1.1.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
tzlocal==2.1
urllib3==1.26.3
validators==0.18.2
Wand==0.6.5
watchdog==1.0.2
wcwidth==0.2.5
WeasyPrint==52.2
webencodings==0.5.1
widgetsnbextension==3.5.1
